Fix string escaping
If a string contains quote-characters, they don't get correctly
escaped, causing syntax errors in the resulting graphviz file.
This quotes the string *twice*: The first quoting makes it a go-quoted
string literal - the second quoting gets unescaped by graphviz.
We then remove the outermost quotes.

As a result, the string appears as a go-quoted string literal in the
output, which seems to be, what we want.